refactor: Convert PathInformation class to domain class (instead of dto class)
* PathInformation is a very prominent and often used class
* We don't want to duplicate this class all over the place
* Serialization is done via a converter anyways
