+++ src/Api.Rest.Dtos.Tests/Data/ConfigurationSerializationTests.cs
@@ -12,13 +12,9 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 {
 	#region usings
 
-	using System;
-	using System.Collections.Generic;
-	using System.IO;
 	using Newtonsoft.Json;
 	using NUnit.Framework;
 	using Zeiss.PiWeb.Api.Definitions;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
 
 	#endregion
