+++ src/Api.Rest.Dtos.Tests/Data/MeasurementSerializationTests.cs
@@ -14,11 +14,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 
 	using System;
 	using System.Collections.Generic;
-	using System.IO;
 	using Newtonsoft.Json;
 	using NUnit.Framework;
-	using Zeiss.PiWeb.Api.Definitions;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
 
 	#endregion
