+++ src/Api.Rest.Dtos.Tests/Data/PathInformationDtoTest.cs
@@ -13,12 +13,12 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 	#region usings
 
 	using NUnit.Framework;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
+	using Zeiss.PiWeb.Api.Contracts;
 
 	#endregion
 
 	[TestFixture]
-	public class PathInformationDtoTest
+	public class PathInformationTest
 	{
 		#region methods
 
@@ -26,8 +26,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_OnePath_IsBelow_AnotherPath()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "P1" ) );
-			var path2 = new PathInformationDto( PathElementDto.Part( "P1" ), PathElementDto.Char( "C" ) );
+			var path1 = new PathInformation( PathElement.Part( "P1" ) );
+			var path2 = new PathInformation( PathElement.Part( "P1" ), PathElement.Char( "C" ) );
 
 			// act, assert
 			Assert.That( path2.IsBelow( path1 ), Is.True );
@@ -38,8 +38,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_SamePath_IsBelow_ReturnsTrue()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Char( "C1" ), PathElementDto.Char( "X" ) );
-			var path2 = new PathInformationDto( PathElementDto.Char( "C1" ), PathElementDto.Char( "X" ) );
+			var path1 = new PathInformation( PathElement.Char( "C1" ), PathElement.Char( "X" ) );
+			var path2 = new PathInformation( PathElement.Char( "C1" ), PathElement.Char( "X" ) );
 
 			// act, assert
 			Assert.That( path1.IsBelow( path2 ), Is.True );
@@ -49,8 +49,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_ComplexPath_IsBelow_AnotherPath()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "P1" ), PathElementDto.Char( "C" ), PathElementDto.Char( "X" ) );
-			var path2 = new PathInformationDto( PathElementDto.Part( "P1" ), PathElementDto.Char( "C" ), PathElementDto.Char( "Y" ) );
+			var path1 = new PathInformation( PathElement.Part( "P1" ), PathElement.Char( "C" ), PathElement.Char( "X" ) );
+			var path2 = new PathInformation( PathElement.Part( "P1" ), PathElement.Char( "C" ), PathElement.Char( "Y" ) );
 
 			// act, assert
 			Assert.That( path1.IsBelow( path2 ), Is.False );
@@ -62,19 +62,19 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_RootPath_SpecialCase_IsHandled()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "P1" ), PathElementDto.Char( "C" ), PathElementDto.Char( "X" ) );
+			var path1 = new PathInformation( PathElement.Part( "P1" ), PathElement.Char( "C" ), PathElement.Char( "X" ) );
 
 			// act, assert
-			Assert.That( path1.IsBelow( PathInformationDto.Root ), Is.True );
-			Assert.That( PathInformationDto.Root.IsBelow( path1 ), Is.False );
-			Assert.That( path1.ParentPath.IsBelow( PathInformationDto.Root ), Is.True );
+			Assert.That( path1.IsBelow( PathInformation.Root ), Is.True );
+			Assert.That( PathInformation.Root.IsBelow( path1 ), Is.False );
+			Assert.That( path1.ParentPath.IsBelow( PathInformation.Root ), Is.True );
 		}
 
 		[Test]
 		public void Test_Equals_SamePaths_ReturnsTrue()
 		{
 			// arrange
-			var path = new PathInformationDto( PathElementDto.Part( "A" ) );
+			var path = new PathInformation( PathElement.Part( "A" ) );
 
 			// act, assert
 			Assert.That( path.Equals( path ), Is.True );
@@ -84,7 +84,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_Equals_PathWithNull_ReturnsFalse()
 		{
 			// arrange
-			var path = new PathInformationDto( PathElementDto.Part( "A" ) );
+			var path = new PathInformation( PathElement.Part( "A" ) );
 
 			// act, assert
 			Assert.That( path.Equals( null ), Is.False );
@@ -94,8 +94,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_Equals_EmptyPaths_ReturnsTrue()
 		{
 			// arrange
-			var path1 = new PathInformationDto();
-			var path2 = new PathInformationDto();
+			var path1 = new PathInformation();
+			var path2 = new PathInformation();
 
 			// act, assert
 			Assert.That( path1.Equals( path2 ), Is.True );
@@ -105,8 +105,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_Equals_PathsWithDifferentLength_ReturnsFalse()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "A" ), PathElementDto.Part( "B" ) );
-			var path2 = new PathInformationDto( PathElementDto.Part( "A" ) );
+			var path1 = new PathInformation( PathElement.Part( "A" ), PathElement.Part( "B" ) );
+			var path2 = new PathInformation( PathElement.Part( "A" ) );
 
 			// act, assert
 			Assert.That( path1.Equals( path2 ), Is.False );
@@ -116,8 +116,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_Equals_PathsNotEqual_ReturnsFalse()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "A" ) );
-			var path2 = new PathInformationDto( PathElementDto.Part( "B" ) );
+			var path1 = new PathInformation( PathElement.Part( "A" ) );
+			var path2 = new PathInformation( PathElement.Part( "B" ) );
 
 			// act, assert
 			Assert.That( path1.Equals( path2 ), Is.False );
@@ -127,8 +127,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_Equals_PathsEqual_ReturnsTrue()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "A" ) );
-			var path2 = new PathInformationDto( PathElementDto.Part( "A" ) );
+			var path1 = new PathInformation( PathElement.Part( "A" ) );
+			var path2 = new PathInformation( PathElement.Part( "A" ) );
 
 			// act, assert
 			Assert.That( path1.Equals( path2 ), Is.True );
@@ -138,8 +138,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_GetHashCode_IsDifferent_When_Paths_Have_Different_Length()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "A" ), PathElementDto.Part( "B" ) );
-			var path2 = new PathInformationDto( PathElementDto.Part( "A" ) );
+			var path1 = new PathInformation( PathElement.Part( "A" ), PathElement.Part( "B" ) );
+			var path2 = new PathInformation( PathElement.Part( "A" ) );
 
 			// act
 			var hashCode1 = path1.GetHashCode();
@@ -153,8 +153,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_GetHashCode_IsDifferent_When_Paths_Are_Not_Equal()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "A" ) );
-			var path2 = new PathInformationDto( PathElementDto.Part( "B" ) );
+			var path1 = new PathInformation( PathElement.Part( "A" ) );
+			var path2 = new PathInformation( PathElement.Part( "B" ) );
 
 			// act
 			var hashCode1 = path1.GetHashCode();
@@ -168,8 +168,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 		public void Test_GetHashCode_IsSame_When_Paths_Are_Equal()
 		{
 			// arrange
-			var path1 = new PathInformationDto( PathElementDto.Part( "A" ) );
-			var path2 = new PathInformationDto( PathElementDto.Part( "A" ) );
+			var path1 = new PathInformation( PathElement.Part( "A" ) );
+			var path2 = new PathInformation( PathElement.Part( "A" ) );
 
 			// act
 			var hashCode1 = path1.GetHashCode();
