+++ src/Api.Rest.Dtos.Tests/Data/ValueSerializationTests.cs
@@ -14,12 +14,9 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 
 	using System;
 	using System.Collections.Generic;
-	using System.IO;
 	using System.Linq;
 	using Newtonsoft.Json;
 	using NUnit.Framework;
-	using Zeiss.PiWeb.Api.Definitions;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
 
 	#endregion
