--- src/Api.Rest.Dtos/Data/InspectionPlanEntityDto.cs
@@ -1,36 +1,0 @@
-﻿#region copyright
-
-/* * * * * * * * * * * * * * * * * * * * * * * * * */
-/* Carl Zeiss IMT (IZfM Dresden)                   */
-/* Softwaresystem PiWeb                            */
-/* (c) Carl Zeiss 2015                             */
-/* * * * * * * * * * * * * * * * * * * * * * * * * */
-
-#endregion
-
-namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
-{
-	#region usings
-
-	using System.Text.Json.Serialization;
-
-	#endregion
-
-	/// <summary>
-	/// Enumeration of possible inspection plan entities.
-	/// </summary>
-	[Newtonsoft.Json.JsonConverter( typeof( Newtonsoft.Json.Converters.StringEnumConverter ) )]
-	[JsonConverter( typeof( JsonStringEnumConverter ) )]
-	public enum InspectionPlanEntityDto : byte
-	{
-		/// <summary>
-		/// The entity is a part.
-		/// </summary>
-		Part = 1,
-
-		/// <summary>
-		/// The entity is a characteristic.
-		/// </summary>
-		Characteristic = 2
-	}
-}
